hide generate image, refactor save project to a service
