+++ src/Project/CreateNewProjectButton.tsx
@@ -12,13 +12,9 @@ import {
 import { useState } from "react";
 import { useAIImageGeneratorStore } from "../Editor/Lyrics/Image/store";
 import CreateNewProjectForm, { DataSource } from "./CreateNewProjectForm";
-import {
-  isProjectExist,
-  loadProjects,
-  saveProject,
-  useProjectStore,
-} from "./store";
+import { isProjectExist, loadProjects, useProjectStore } from "./store";
 import { ProjectDetail } from "./types";
+import { useProjectService } from "./useProjectService";
 
 enum CreateProjectOutcome {
   missingStreamUrl = "Missing stream url",
@@ -28,6 +24,7 @@ enum CreateProjectOutcome {
 }
 
 export default function CreateNewProjectButton() {
+  const [saveProject] = useProjectService();
   const [creatingProject, setCreatingProject] =
     useState<ProjectDetail | undefined>();
   const [selectedDataSource, setSelectedDataSource] = useState<DataSource>(
@@ -44,12 +41,12 @@ export default function CreateNewProjectButton() {
     (state) => state.setUnsavedLyricReference
   );
   const setLyricReference = useProjectStore((state) => state.setLyricReference);
-  
+
   const setPromptLog = useAIImageGeneratorStore((state) => state.setPromptLog);
   const setGeneratedImageLog = useAIImageGeneratorStore(
     (state) => state.setGeneratedImageLog
   );
-  
+
   const [attemptToCreateFailed, setAttemptToCreateFailed] =
     useState<boolean>(false);
   const [createProjectOutcome, setCreateProjectOutcome] =
@@ -77,8 +74,8 @@ export default function CreateNewProjectButton() {
         setLyricTexts([]);
         setUnSavedLyricReference("");
         setLyricReference("");
-        setPromptLog([])
-        setGeneratedImageLog([])
+        setPromptLog([]);
+        setGeneratedImageLog([]);
 
         close();
         setCreatingProject(undefined);
