+++ src/Project/SaveButton.tsx
@@ -1,32 +1,13 @@
-import { ActionButton, Button, Text } from "@adobe/react-spectrum";
-import { useAIImageGeneratorStore } from "../Editor/Lyrics/Image/store";
-import { saveProject, useProjectStore } from "./store";
+import { Button, Text } from "@adobe/react-spectrum";
+import { useProjectService } from "./useProjectService";
 
 export default function SaveButton() {
-  const editingProject = useProjectStore((state) => state.editingProject);
-  const lyricTexts = useProjectStore((state) => state.lyricTexts);
-  const unSavedLyricReference = useProjectStore(
-    (state) => state.unSavedLyricReference
-  );
-  const generatedImageLog = useAIImageGeneratorStore(
-    (state) => state.generatedImageLog
-  );
-  const promptLog = useAIImageGeneratorStore((state) => state.promptLog);
-
+  const [saveProject] = useProjectService();
   return (
     <Button
       variant={"primary"}
       onPress={() => {
-        if (editingProject) {
-          saveProject({
-            id: editingProject.name,
-            projectDetail: editingProject,
-            lyricTexts,
-            lyricReference: unSavedLyricReference,
-            generatedImageLog,
-            promptLog,
-          });
-        }
+        saveProject();
       }}
     >
       <Text>Save</Text>
